Move sharing function to separate js file
